Add files via upload
Map-Reduce Program
In this project, you are asked to implement a simple graph algorithm that needs two Map-Reduce jobs. A directed graph is represented as a text file where each line represents a graph edge. For example,
20,40
represents the directed edge from node 20 to node 40. First, for each graph node, you compute the number of node neighbors. Then, you group the nodes by their number of neighbors, and for each group, you count how many nodes belong to this group. That is, the result will have lines such as:
10 30
which says that there are 30 nodes that have 10 neighbors.
You should write the two Map-Reduce job in the Java file.

In your Java main program, args[0] is the graph file, and args[1] is the output directory. The input file format for reading the input graph and the output format for the final result must be text formats, while the format for the intermediate results between the Map-Reduce jobs must be binary formats